requirements.txt
streamlit
tensorflow==2.21.0
numpy
pillow
h5py